fix(natsutil): move real-NATS RequestFailure test to embedded-server unit test
The Docker-backed integration test could never run in Docker-less
sandboxes/CI and duplicated coverage the package's existing embedded NATS
server (startTestNATSWithMaxPayload) already provides for free. Replace
it with a unit test against that in-process server, so the no-responders
path is verified by every `make test` run instead of being skipped
wherever a Docker daemon is absent.

Moved request_failure_test.go from `package natsutil` to `package
natsutil_test` (matching reply_test.go's package, where the embedded-
server helper lives) and qualified calls as natsutil.RequestFailure —
required because the helper is unexported and the two test packages in
this directory are compiled separately; the brief's original snippet
assumed same-package visibility that doesn't hold here.
